requirements.txt
Flask==3.0.0
twilio==8.10.0
openai==1.3.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
numpy>=2.0.0
gunicorn==21.2.0
